requirements.txt
streamlit==1.25.0
rembg==2.0.50
onnxruntime==1.16.3
Pillow==9.4.0
matplotlib==3.7.1
tensorflow==2.13.0
numpy
psycopg2-binary
pydrive2
Pillow
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
